restructure view modules
