+++ src/global/message.rs
@@ -4,7 +4,7 @@ use ratatui_image::protocol::StatefulProtocol;
 use rust_ffmpeg::FFmpegProcess;
 
 use crate::{
-    global::logic::{confirmation::Response, terminal::CursorMovementDirection},
+    global::{logic::confirmation::Response, view_logic::terminal::CursorMovementDirection},
     user_input::logic::{InputTarget, PromptMsg},
 };
