+++ src/global/mod.rs
@@ -3,3 +3,4 @@ pub mod logic;
 pub mod message;
 pub mod update;
 pub mod view;
+pub mod view_logic;
