+++ src/global/view/info_popup.rs
@@ -7,7 +7,7 @@ use ratatui::{
 
 use crate::{global::view::center_xy, model::Model};
 
-pub fn render(model: &Model, frame: &mut Frame, area: Rect) {
+pub fn draw(model: &Model, frame: &mut Frame, area: Rect) {
     let popup_block = Block::bordered()
         .title("Info ")
         .title_alignment(Alignment::Left)
