+++ src/global/view_logic/keybinds.rs
@@ -1,65 +1,19 @@
-use ratatui::{
-    Terminal,
-    backend::CrosstermBackend,
-    crossterm::{
-        ExecutableCommand,
-        terminal::{EnterAlternateScreen, LeaveAlternateScreen, disable_raw_mode, enable_raw_mode},
-    },
-};
-use std::{
-    io::{Stdout, stdout},
-    panic,
-};
-
-use std::time::Duration;
-
-use color_eyre::eyre::Result;
-use crossterm::event::{self, Event, KeyCode, KeyEvent, KeyEventKind};
+use crossterm::event::{KeyCode, KeyEvent};
 
 use crate::{
     global::{
         logic::confirmation::Response,
         message::{Message, PlayerMessage, PlaylistMessage, UserInputMessage},
-        view::tabs::SelectedTab,
+        view_logic::{focused_area::FocusedArea, terminal::CursorMovementDirection},
     },
     model::Model,
     user_input::logic::InputMode,
 };
 
-#[derive(Clone, Copy)]
-pub enum CursorMovementDirection {
-    Up,
-    Down,
-    Left,
-    Right,
-}
-
-pub fn init_terminal() -> color_eyre::Result<Terminal<CrosstermBackend<Stdout>>> {
-    enable_raw_mode()?;
-    stdout().execute(EnterAlternateScreen)?;
-    let terminal = Terminal::new(CrosstermBackend::new(stdout()))?;
-    Ok(terminal)
-}
-
-pub fn restore_terminal() -> color_eyre::Result<()> {
-    stdout().execute(LeaveAlternateScreen)?;
-    disable_raw_mode()?;
-    Ok(())
-}
-
-pub fn install_panic_hook() {
-    let original_hook = panic::take_hook();
-    panic::set_hook(Box::new(move |panic_info| {
-        stdout().execute(LeaveAlternateScreen).unwrap();
-        disable_raw_mode().unwrap();
-        original_hook(panic_info);
-    }));
-}
-
-fn handle_keys(key_event: KeyEvent, model: &Model) -> Option<Message> {
+pub fn handle_key_inputs(key_event: KeyEvent, model: &Model) -> Option<Message> {
     match model.input_mode {
         InputMode::Commands => match model.selected_tab {
-            SelectedTab::Main => match key_event.code {
+            FocusedArea::Main => match key_event.code {
                 KeyCode::Esc => Some(Message::Quit),
                 KeyCode::Char('n') => Some(Message::Player(PlayerMessage::LoadNow)),
                 KeyCode::Char(' ') => Some(Message::Player(PlayerMessage::TogglePlay)),
@@ -80,7 +34,7 @@ fn handle_keys(key_event: KeyEvent, model: &Model) -> Option<Message> {
                 _ => None,
             },
 
-            SelectedTab::Playlist => match key_event.code {
+            FocusedArea::Playlist => match key_event.code {
                 KeyCode::Esc => Some(Message::Quit),
                 KeyCode::Char('c') => Some(Message::Playlist(PlaylistMessage::ToggleControlPanel)),
                 KeyCode::Char('a') => Some(Message::Playlist(PlaylistMessage::ToggleArrangeTracks)),
@@ -126,16 +80,4 @@ fn handle_keys(key_event: KeyEvent, model: &Model) -> Option<Message> {
         },
         InputMode::Info => Some(Message::AcknowledgeInfo),
     }
-}
-
-pub fn handle_events(model: &Model) -> Result<Option<Message>> {
-    if event::poll(Duration::from_millis(16))? {
-        match event::read()? {
-            Event::Key(key_event) if key_event.kind == KeyEventKind::Press => {
-                return Ok(handle_keys(key_event, model));
-            }
-            _ => {}
-        };
-    }
-    Ok(None)
 }
