+++ src/global/view_logic/mod.rs
@@ -1,0 +1,3 @@
+pub mod focused_area;
+pub mod keybinds;
+pub mod terminal;
