+++ src/player/view/control_bar.rs
